Feature thymeleaf CRUD

+++ src/main/java/raisetech/student/management/data/Student.java
@@ -12,7 +12,7 @@ public class Student {
   private String nickname;
   private String email;
   private String area;
-  private int age;
+  private short age;
   private String gender;
   private String remark;
   private boolean isDeleted;
